Refactor SQLAlchemy + Marshmallow integration
Hopefully both working and using Flask extension style (tests to follow)
