+++ .travis.yml
@@ -11,9 +11,9 @@ matrix:
 env:
   - APP_CONFIG_NAME=test
 
-script:  # Handles creating a blank config file so imports dont fail
+script:
   - pytest
-  - flake8 middleware/ --ignore=F403,F405
+  - flake8 middleware/ --ignore=F401,F403,F405
 install:
  - pip install -r requirements.txt
