--- middleware/core.py
@@ -1,2 +1,0 @@
-from flask_sqlalchemy import SQLAlchemy
-db = SQLAlchemy()
