+++ requirements.txt
@@ -6,13 +6,16 @@ cffi==1.10.0
 click==6.7
 cryptography==2.0.2
 Flask==0.12.2
+flask-marshmallow==0.8.0
 Flask-RESTful==0.3.6
+Flask-SQLAlchemy==2.2
 idna==2.5
 itsdangerous==0.24
 Jinja2==2.9.6
 json-merge-patch==0.1
 Mako==1.0.7
 MarkupSafe==1.0
+marshmallow-sqlalchemy==0.13.1
 paramiko==2.2.1
 py==1.4.34
 pyasn1==0.3.1
